Bug 39507: Make the MARC21 008 plugin more precise for MU
According to https://loc.gov/marc/bibliographic/bd008m.html, the content
of 008/24-29 and 008/30-31 fr Music (leader/6 = c, d, i, or j) is made
of one character codes (as it is the case for 008/24-27 for Books).
Now, in Koha 008/24-29 and 008/30-31 are input fields with no
possibility to chose the code from a list. This could (and should)
be improved.

Test plan:
==========
1. Open a standard biblio editor, chose the plugin for 008 field,
   change Type of Material to MU - Music. See that 24-29 and 30-31
   (Accompanying matter and Literary text for sound recordings )are
   multi character input fields.
2. Apply the patch ; restart_all.
3. Repeat p. 1. See that now 24-29 and 30-31 are split in a series
   of one-character drop-down menus.

BTW a link to a list of codes for 'Form of composition' (008/18-19)
has been also added.

Signed-off-by: Roman Dolny <roman.dolny@jezuici.pl>
Signed-off-by: Tomás Cohen Arazi <tomascohen@theke.io>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>