Bug 39567: Include form-submit code from opac-bottom.inc (OPAC)
Import the form-submit.js script in the global includes file for the
OPAC, so that all pages have access to it without needing to
import it directly.

This patch also aligns the OPAC copy of form-submit.js to the staff
interface copy, as the OPAC copy is missing 3 lines to support opening
the link in a new tab (added to the staff interface copy in bug 37192).

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@openfifth.co.uk>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>