Bug 39739: Sync OPAC's datatables.js with the staff interface's version
This patch overwrites the OPAC version of `datatables,js` with the one
for the staff interface.

A block code adding links to the columns configuration page is removed
as it doesn't apply to the OPAC and produces errors in the console for
missing variable definitions that would be inherited from the caller
context (looking for `CAN_user_parameters_manage_column_config`).

To test:
1. Apply this patch
2. Find templates using `kohaTable`:
   $ cd koha-tmpl/opac-tmpl
   $ git grep kohaTable
3. Play in the UI with those pages
=> SUCCESS: Tables work, nothing's broken
4. Sign off :-D

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
Signed-off-by: Martin Renvoize <martin.renvoize@openfifth.co.uk>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>