Bug 39779: (QA follow-up) Undo revert and instead call .api()
This patch leaves Laura's patch intact for the record, but effectively
reverts it's reversion of bug 38255 in which all ajax DataTables were
converted to a standard kohaTable instantiation structure.

We need a DataTable() api in subsequent uses of the table and not the
jQuery object that kohaTable returns so I update all calls to
saleDataTable to sale_kohaTable.api() which returns us the DataTable()
api instance from the dataTable jQuery object.

Signed-off-by: Martin Renvoize <martin.renvoize@openfifth.co.uk>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>