Bug 30200: (QA follow-up) Keep multiselect in sync
This patch updates the multiselect status picker handling to ensure it
remains in sync with the active tab.

On tab change we now disable statuses that are not part of the tab
select options, deselect any other status filters and then select all
tab statuses.  We allow further refinement of the filters still by
allowing the user to deselect any of the available statuses. The filters
should continue to work as before except using the 'Clear' button will
now reapply the tab defined status set in full as apposed to removing
all status filtering entirely.

Signed-off-by: Martin Renvoize <martin.renvoize@openfifth.co.uk>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>