Bug 39054: a need to distinguish between manual bibliographic record modifications and those being a consequence of authority modification
Now, in action_logs, Koha does not distinguish between
manual bibliographic record modifications and those being
a consequence of authority modification. Both are logged
in the same way: module: CATALOGUING ; operation: MODIFY
; interface: intranet.

It will be helpful to be able to distinguish this two
cases (for instance to be able to evaluate a librarian's
'koha' in the cataloguing process: a single  action of
modification of one authority record launching a
modification of, say, 100 bibliographic records does not
equal to manual modification of 100 bibliographic
records. Also, if we trace issues in bibliographic
records after modification being performed and want
librarians to correct themselves, a biblio modification
being a consequence of an authority modification should
not be taken into account, etc.).

Test plan:
==========
1. Have CataloguingLog system preference set to Log.
2. Find any bibliographic record linked with a authority
   record (e.g. sn:1 linked to Heylin, Clinton).
3. Modify the bibliographic record (Edit > Edit record >
   [make an edit] > Save).
3. In a separate tab, open the authority record and
   modify the heading.
4. Go back to the bibliographic record tab and chose
   Modification log. You will see the two recent operations
   marked as 'Modify'. You are not able to distinguish
   between the 'real' manual modification from the second
   being the result of an authority record modification.
5. Apply the patch ; restart all.
6. Repeat p. 3 and 4. You should now see the last action
   (resulting from authority modification) marked as
   'Merge', which allows you to distinguish the two
   different actions.

Signed-off-by: Roman Dolny <roman.dolny@jezuici.pl>